requirements.txt
beautifulsoup4==4.12.3
html5lib==1.1

httpx==0.27.2

requests==2.32.3
loguru==0.7.3
lxml==5.3.0
tenacity==9.0.0
python-dotenv==1.0.1
selenium==4.27.1
APScheduler==3.11.0
pytz==2025.1

# gunicorn==23.0.0

# Приложение и работа с jwt
fastapi==0.115.12
uvicorn==0.34.2
# python-jose==3.3.0


# ORM, миграции и работа с базой
tortoise-orm==0.24.0
asyncpg==0.30.0
aerich==0.7.2